requirements.txt
dash==2.3.0
pandas==1.4.1
plotly-express==0.4.1